Upgrading target OS to Centos 7 x86_64.
I want to use a 32-bit OS, but want to use the latest version even more.
I got all of the modules working with minimal changes with the exception of the Hadoop module.
The Hadoop module is temporarily disabled for now, but should be usable one I recreate the rpm for x86_64.