Lots of changes.
Adding Zookeeper support.
There is a finicky setup with the myid files, but it's good enough for now.

Adding secondarynamenode as a managed daemon.
I completely forgot that was a thing.

Removed all *.cmd files.
Updated hadoop-env.sh and yarn-env.sh to have sensible defaults for JAVA_HOME.

Updated the configs in puppet to better match what comes with Hadoop 2.5.1.
Random cleanup of manifests.
Adding a directory for YARN's local NodeManager stuff.