Big admin changes here!
The daemons will no longer be configured to run at machine startup.
The cross machine dependencies were causing too many errors.

Replaced the systemd startup with a simple generated set of scripts
that start and stop the daemons in a ordered and predictable fashion.
I still use systemd as the daemon manager.

This will also allows me to configure what daemons I want to startup.

Modified the admin_user sudo file to allow for passwordless
systemctl operations on the daemons managed in this repo.
This combined with the passwordless ssh makes starting and stopping
the grid easy as pie.

Expanded the hieradata files to support for specifying your hbase
and solr nodes.